feat: Add confirmation dialogs (#2253)
- RSS Feed deletion
- RSS Rule deletion
- Peer ban
- tag / category deletion
- Right Click > remove unused tags / categories

+++ src/components/RSS/Feeds/Feeds.vue
@@ -1,5 +1,7 @@
 <script setup lang="ts">
+import ConfirmDialog from '@/components/Dialogs/Confirm/ConfirmDialog.vue'
 import RssFeedDialog from '@/components/Dialogs/RssFeedDialog.vue'
+import { useI18nUtils } from '@/composables'
 import { useDialogStore, useRssStore } from '@/stores'
 import { Feed } from '@/types/qbit/models'
 import { RssArticle } from '@/types/vuetorrent'
@@ -17,6 +19,7 @@ defineEmits<{
   openArticle: [article: RssArticle]
 }>()
 
+const { t } = useI18nUtils()
 const dialogStore = useDialogStore()
 const rssStore = useRssStore()
 
@@ -47,8 +50,15 @@ async function refreshFeed(item: Feed) {
 }
 
 async function deleteFeed(item: Feed) {
-  await rssStore.deleteFeed(item.name)
-  rssStore.resumeFeedTimer()
+  dialogStore.createDialog(ConfirmDialog, {
+    title: t('dialogs.confirm.deleteFeed'),
+    text: item.name,
+    yesColor: 'error',
+    onConfirm: async () => {
+      await rssStore.deleteFeed(item.name)
+      rssStore.resumeFeedTimer()
+    }
+  })
 }
 
 onMounted(() => {
