+++ src/components/Settings/TagsAndCategories.vue
@@ -1,22 +1,62 @@
 <script setup lang="ts">
 import CategoryFormDialog from '@/components/Dialogs/CategoryFormDialog.vue'
+import ConfirmDialog from '@/components/Dialogs/Confirm/ConfirmDialog.vue'
+import ConfirmListDialog from '@/components/Dialogs/Confirm/ConfirmListDialog.vue'
 import TagFormDialog from '@/components/Dialogs/TagFormDialog.vue'
+import { useI18nUtils } from '@/composables'
 import { useCategoryStore, useDialogStore, useMaindataStore, useTagStore } from '@/stores'
 import { Category } from '@/types/qbit/models'
 
+const { t } = useI18nUtils()
 const categoryStore = useCategoryStore()
 const dialogStore = useDialogStore()
 const maindataStore = useMaindataStore()
 const tagStore = useTagStore()
 
 async function deleteTag(tagName: string) {
-  await tagStore.deleteTags([tagName])
-  maindataStore.forceMaindataSync()
+  dialogStore.createDialog(ConfirmDialog, {
+    title: t('dialogs.confirm.deleteTag'),
+    text: tagName,
+    yesColor: 'error',
+    onConfirm: async () => {
+      await tagStore.deleteTags([tagName])
+      maindataStore.forceMaindataSync()
+    }
+  })
+}
+
+async function deleteUnusedTags() {
+  dialogStore.createDialog(ConfirmListDialog, {
+    title: t('dialogs.confirm.deleteUnusedTags'),
+    items: tagStore.unusedTags,
+    yesColor: 'error',
+    onConfirm: async () => {
+      await tagStore.deleteUnusedTags()
+    }
+  })
 }
 
 async function deleteCategory(category: Category) {
-  await categoryStore.deleteCategories([category.name])
-  maindataStore.forceMaindataSync()
+  dialogStore.createDialog(ConfirmDialog, {
+    title: t('dialogs.confirm.deleteTag'),
+    text: category.name,
+    yesColor: 'error',
+    onConfirm: async () => {
+      await categoryStore.deleteCategories([category.name])
+      maindataStore.forceMaindataSync()
+    }
+  })
+}
+
+async function deleteUnusedCategories() {
+  dialogStore.createDialog(ConfirmListDialog, {
+    title: t('dialogs.confirm.deleteUnusedCategories'),
+    items: categoryStore.unusedCategories,
+    yesColor: 'error',
+    onConfirm: async () => {
+      await categoryStore.deleteUnusedCategories()
+    }
+  })
 }
 
 function openTagFormDialog(initialTag?: string) {
@@ -32,7 +72,7 @@ function openCategoryFormDialog(initialCategory?: Category) {
   <v-row>
     <!-- Tags -->
     <v-col cols="12" sm="6">
-      <v-list-subheader class="ml-2">{{ $t('settings.tagsAndCategories.tagsSubheader') }}</v-list-subheader>
+      <v-list-subheader class="ml-2">{{ t('settings.tagsAndCategories.tagsSubheader') }}</v-list-subheader>
 
       <v-sheet rounded="xl" class="d-flex align-center gap" v-for="tag in tagStore.tags">
         <div class="pl-4 py-1 wrap-anywhere">{{ tag }} ({{ tagStore.torrentsByTag[tag] ?? 0 }})</div>
@@ -44,22 +84,22 @@ function openCategoryFormDialog(initialCategory?: Category) {
       </v-sheet>
 
       <v-card v-if="tagStore.tags.length === 0">
-        <v-card-text>{{ $t('settings.tagsAndCategories.noTags') }}</v-card-text>
+        <v-card-text>{{ t('settings.tagsAndCategories.noTags') }}</v-card-text>
       </v-card>
 
       <div class="d-flex ga-2">
         <v-btn class="flex-grow-1" color="accent" prepend-icon="mdi-plus" @click="openTagFormDialog()">
-          {{ $t('settings.tagsAndCategories.createNewTag') }}
+          {{ t('settings.tagsAndCategories.createNewTag') }}
         </v-btn>
-        <v-btn class="flex-grow-1" color="error" prepend-icon="mdi-delete" @click="tagStore.deleteUnusedTags()">
-          {{ $t('settings.tagsAndCategories.deleteUnusedTags') }}
+        <v-btn class="flex-grow-1" color="error" prepend-icon="mdi-delete" :disabled="tagStore.unusedTags.length === 0" @click="deleteUnusedTags()">
+          {{ t('settings.tagsAndCategories.deleteUnusedTags') }}
         </v-btn>
       </div>
     </v-col>
 
     <!-- Categories -->
     <v-col cols="12" sm="6">
-      <v-list-subheader class="ml-2">{{ $t('settings.tagsAndCategories.categoriesSubheader') }}</v-list-subheader>
+      <v-list-subheader class="ml-2">{{ t('settings.tagsAndCategories.categoriesSubheader') }}</v-list-subheader>
 
       <v-sheet rounded="xl" class="d-flex align-center gap" v-for="category in categoryStore.categories">
         <div class="pl-4 py-1 wrap-anywhere">{{ category.name }} ({{ categoryStore.torrentsByCategory[category.name] ?? 0 }})</div>
@@ -71,15 +111,15 @@ function openCategoryFormDialog(initialCategory?: Category) {
       </v-sheet>
 
       <v-card v-if="categoryStore.categories.length === 0">
-        <v-card-text>{{ $t('settings.tagsAndCategories.noCategories') }}</v-card-text>
+        <v-card-text>{{ t('settings.tagsAndCategories.noCategories') }}</v-card-text>
       </v-card>
 
       <div class="d-flex ga-2">
         <v-btn class="flex-grow-1" color="accent" prepend-icon="mdi-plus" @click="openCategoryFormDialog()">
-          {{ $t('settings.tagsAndCategories.createNewCategory') }}
+          {{ t('settings.tagsAndCategories.createNewCategory') }}
         </v-btn>
-        <v-btn class="flex-grow-1" color="error" prepend-icon="mdi-delete" @click="categoryStore.deleteUnusedCategories()">
-          {{ $t('settings.tagsAndCategories.deleteUnusedCategories') }}
+        <v-btn class="flex-grow-1" color="error" prepend-icon="mdi-delete" :disabled="categoryStore.unusedCategories.length === 0" @click="deleteUnusedCategories()">
+          {{ t('settings.tagsAndCategories.deleteUnusedCategories') }}
         </v-btn>
       </div>
     </v-col>
